Fix regex, remove awk and xdg-mime dependency
The [] operator does not group characters, but it matches any character
in the group, i.e. [https://doi.org/] would match just a single /.  To
actually match the optional substring you need parentheses (only works
with extended regexp).

xdg-mime is unnecessary to check for a PDF file.  The PDF standard
demands that the first line of a PDF file always has the form

    %PDF-1.4

where 1.4 is version and can vary.  So it's easier to just grab the
first four bytes of a file and check whether it's "%PDF".

The awk dependency is unnecessary as well because stripping "Title:" can
be done with pure shell substitution.  To look up whether the DOI
already exists in the $BIB file, awk is actually harmful because the DOI
always contains a slash which breaks regexp.  Instead it's probably
better to just search for the fixed string of the DOI in the file.
That's more robust and since DOIs are unique you should not have two
different bib entries with the same DOI.